fix: #26 完成释放技能时的逻辑顺序调整
skillevent 的逻辑顺序已调整正确.
